refactor: rpc client/hook
