+++ frontend/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './rpc';
 export * from './inputs';
 export * from './editor';
 export * from './store';
